refactor: utils.js -> characterData.js 변경 (#30)
