+++ developer-test-template/src/lib/characterData.js
